Get different sections/parts from ajax requests
Instead of adding the markup for different components such as the
authentication form, console, presence buttons etc, these now come from
ajax requests and are the actions on the buttons are binded in the
respective callbacks. This will make all functionalities 'composable'.

For eg, the presence buttons section comes from /bosh/presencePanel.html
and in the callback the respective buttons are bound