Refactor and rename directories and files
All html files are now moved inside the 'html' dir from the 'bosh'
dir. All files that are snippets to be obtained via ajax are prefixed
with an underscore (eg. _authForm.html). The name of the files to be
opened in browser stay the same (eg. connect.html).

The references in javascript are also changed respectively.